Fixed duplicate functions and improved modularity of activities
